Fix Hiding When Busy (#3987)
# About the pull request

This PR makes it so the xenohide ability requires the xeno to not be
busy (such as channeling a pounce). #3902 almost addressed this, but not
in the scenario of not being hidden -> pounce start -> insta hide.

# Explain why it's good for the game

Fixes working around unhide mechanics like this (can be much faster than
my example but I just had hide bound to a separate key):
https://cdn.discordapp.com/attachments/745447048261795890/1133194895918649354/hug.mp4

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

See above.

</details>


# Changelog
:cl: Drathek
fix: Fixed xeno hide ability not checking for busy status.
/:cl: