Ares Nuke Request (#3988)
# About the pull request

Implements ARES Emergency Procedure 'Request Nuke', which allows someone
to petition High Command (Admins) for a nuke

# Explain why it's good for the game

This is already implemented in the framework for Project ARES, this just
adds the feature. Should be fine as admins will need to authorize it. If
people want a different rank for such nuke requests I can have it done.

# Changelog
:cl:
add: You can now petition High Command via ARES for a nuclear device,
bypassing techweb (If you can convince them, that is.)
admin: Implemented Admin side of ARES Nuke request to allow for granting
of different nuke types, or denying the request.
/:cl:

---------

Co-authored-by: forest2001 <ingram.josh@gmail.com>
Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>