Reworks facehugged effects to make it harder to ignore (#3843)
# About the pull request

This makes the paralyze chance when facehugged apply earlier in the
hugged phase, and it has a higher chance of happening, but paralyzes for
less time when it does happen. Paralyze chance gets slightly higher with
a slightly longer time the longer you stay hugged. Changes paralyze
message slightly to clue new players into seeking help.

# Explain why it's good for the game

When you're hugged, it should be scary and urgent to deal with, to make
the game more interesting and time-sensitive - rather than "*yawn* time
for surgery in awhile, I guess, after a bit more fighting." I know
balance changes are controversial and I'm very open to changing this PR,
but I wanted to at least start a conversation by opening this. My goal
is not to nerf marines (if marines get buffed in some other way in a
different PR, great) I just want to make the game more interesting and
fun.


# Changelog
:cl:
balance: when facehugged, more paralyze chance with a smaller paralyze
duration
ui: during hugged effects, changes messaging to encourage new players to
seek help
/:cl:

---------

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>