Another Banishment Check For Observers Joining as a Xeno (#4039)
# About the pull request

This PR is a followup to #4031 where I found that ctrl clicking an afk
xeno is also a method to join as a xeno. Now this method also checks for
banishment.

# Explain why it's good for the game

Rejoining as a different mob should not circumvent banishment.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/2bed0b6a-b3e6-498b-a7f7-c18bfd48eb6e)

</details>


# Changelog
:cl: Drathek
fix: Fixed ctrl+click joining as an afk xeno not checking banishment
/:cl: