Unnested Larva Maturation Prompts (#4041)
# About the pull request

This PR adds 10s tgui_alerts asking an observer to take over a maturing
larva (an embryo that reached stage 5 in a host) if the host is
unnested. They will automatically orbit the bursting mob when prompted.
It will make 5 attempts at this. If the host is nested, the process will
proceed as it has without any prompts.

This PR also fixes the pool flash not occurring for bursts from hosts.

# Explain why it's good for the game

For the pool flash, that should have just been occurring in the first
place.

For the prompts:
- This allows observers to opt out of accepting a larva spawn in what
they deem an impossible scenario - or at least a scenario they don't
want to risk their place in queue for.
- This now makes up to 5 attempts from the queue (technically 6 if you
count a player hugger 2nd dibs) to put a player into a bursting larva
rather than a single attempt. A side effect to this though is it can
potentially increase the time to autoburst up to one minute while we're
querying candidates.
- This ensures that the new larva has an active player that didn't step
away a moment ago (the queue allows a 5 minute window for inactivity)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl: Drathek
add: Added prompts for xeno candidates to optionally take over a larva
when an unnested host is ready to burst
fix: Fixed larva spawning in hosts not triggering the preference unpool
taskbar flash.
/:cl: