Standardize hugger attack windup and attack windup duration (#4044)
# About the pull request

This PR sets the attack windup duration for a player facehugger to 1
second, like it is for a pounce.

# Explain why it's good for the game

Since you can just pounce at point blank, there is zero reason for a
point blank attack (that also requires them tackled) to take 6x longer.
Any hugger that would do this would obviously be new to the caste for
doing it in an inferior way.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/cmss13-devs/cmss13/assets/76988376/c81c1635-42a9-4b6a-8411-e99b562fdb98

</details>


# Changelog
:cl: Drathek
balance: Facehugger attack windup duration is now 1s like the pounce
windup.
/:cl: