fix having inapropriate ammo in CL lockbox. (#4071)
# About the pull request
fixes: https://github.com/cmss13-devs/cmss13/issues/3976
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
removing the improper ammo in the lock box because it's obsolete
considering the new gun CL as.
adding a way to buy more ammo so that CL as a way to refill is weapon.
didn't wanted to give it for free in the lockbox.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/117036822/f9aeb4e9-e11f-4a3f-9ccd-da244ee74713)


![image](https://github.com/cmss13-devs/cmss13/assets/117036822/815b6f6a-0532-4beb-9a68-7c524ead3eb1)


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: adding the possibility to buy "ES-4 stun magazine" for 10 point
in the portable vendor.
fix: remove a bunch of ammo that where improper for ES-4 gun that are in
the lockbox.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>