Lesser Drone now Has 3 Guarantied Slots (#4045)
# About the pull request

Lesser Drones now will always have at least 3 slots, no matter the state
of the hive

# Explain why it's good for the game

Lesser Drones as supposed to be a swarm, and at this stage of the game
many Xenos are likely ghosting. This gives them a bigger chance to
participate in rounds (at least the ones where Xenos are on the
backfoot). This should make it so lesser drones have a baseline of 3
slots, plus additional limit based on hive size.

# Changelog
:cl:
add: Lesser Drone Slots now have a minimum, currently set to 3.
/:cl: