J.O.E.Y. [Hyperstyled Edition] (#4042)
# About the pull request

Replaces the cigarette vendor in maint bar with the workin' joey prop

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

_**V1BRANT S0VL**_


![image](https://github.com/cmss13-devs/cmss13/assets/103964974/d0d201a3-24a0-4a39-8353-a400f6f9987a)
# Changelog
:cl: Thwomper
maptweak: added Workin' Joey to maint bar.
/:cl:

---------

Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>