BUG FIX: 3rd Party Music Filename References (#4082)
# About the pull request
3rd Party victory music isn't playing when WY/UPP/CLF win a round, I
believe this is why:

Was told to update the .ogg files to all lower case in the original PR.
Unfortunately when I referenced them in code they still had
capitalization, and that was not caught. This should fix 3rd Party music
not playing. (I think).

# Explain why it's good for the game
Bug fixes good


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: fixed a file reference that broke end of round music for some
factions
/:cl: