Workin Joey Talks (#4093)
# About the pull request

Expansion on #4042 that was discussed with Thwomp. Working Joe randomly
says a quip and responds when attacked.

# Explain why it's good for the game

More character to Joey instead of being a static prop that doesn't move.
Showing how "Joey" is still active but been interfered with by the crew.

# Testing Photographs and Procedure


![image](https://github.com/cmss13-devs/cmss13/assets/91219575/f24d5095-301d-4d2c-896b-3134928d7e43)

# Changelog
:cl:
add: Workin Joey now randomly talks and responds when attacked
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>