Falcon Drone appears on marine MD (#4057)
# About the pull request
Adds the falcon drone to marine MD, same manner as queen eye.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
It is a moving object after all, more so than the Queen's eye.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added the falcon drone to marine motion detectors, appearing the
same as Queen eye.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>