Preds now show up on the predship tacmap (#4080)
# About the pull request
Now makes preds (and thralls) show up on their (and only their) tacmaps.

Additionally, anyone wearing a stolen bracer (since that's what the
tracker is based from) has a custom icon on the tacmap.

# Explain why it's good for the game
Well, it makes sense that preds should be able to see each other on
their tacmap. Not sure if this was intended when preds got their tacmap,
but it's here now.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Predators and thralls are now visible on the predship tacmap.
add: Anyone wearing a stolen predator bracer is now visible on the
predship tacmap.
/:cl:

---------

Co-authored-by: John Doe <johndoe@gmail.com>