You can no longer bypass skill checks on pill bottles (#4097)
# About the pull request

When it got passed to attack_hand() from unarmed_attack() in click() it
would not check the skillcheck as it calls attack_hand() on the pill
rather than through the pill bottle which was allowing a bypass for the
skillcheck for pill bottles. This *feels* like the wrong way to do it as
like OOP so I'm happy to do it in a different way.

# Explain why it's good for the game

Bug bad


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: You can no longer bypass skill checks on pill bottles
/:cl: