Adds a False Testemony charge to Marine Law (#4009)
# About the pull request

Adds a new Marine Law charge "False Testemony" to the JAS system. It was
classified as a minor crime and 5 minutes brig timer.

# Explain why it's good for the game

Right now, Marine Law doesn't cover people purposely lying to MPs when
they are asked about whether they saw who did what, etc. The closest
thing we have is aiding and abetting but that usually only applies when
someone actively attempts to help a suspect.

If the person that lied to the MP is proved to have lied, Marine Law
should cover this circumstance.


# Testing Photographs and Procedure




![image](https://github.com/cmss13-devs/cmss13/assets/52742585/751642c0-88a7-4ec7-a553-1487b3650a33)


# Changelog


:cl: smov
add: Adds a new Marine Law charge
/:cl:

---------

Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>