Admin Flag for Disposal mobs (#3814)
# About the pull request
Allows admins to toggle a flag as to whether mobs will fit inside
disposals tubes.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Regarding disposal tubes, there are some situations where I can see it
being fitting to the situation, and this will also allow a measured
level of return for a feature we had to remove due to abuse.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added an admin toggle flag for fitting mobs into disposals. (Within
reason - no queens...)
add: The Crawler trait that allows mobs to enter vents will also allow
them to climb into disposals regardless of above flag.
add: Added a subtype of disposal that can't fit mobs, used for the
delivery units.
/:cl: