Nightmare Insert: Trijent Dam - Crashed UPP Survivors (#3845)
# About the pull request

**Starring**
_The UPP 173rd Airborne Reconnaissance Company, 2nd Platoon_
![Screenshot 2023-07-09 16 10
32](https://github.com/cmss13-devs/cmss13/assets/6595389/76159e01-6d7f-44f7-b404-e30c54ecc9c8)

This PR adds a new nightmare insert to Trijent Dam that sees a small
group of UPP survivors, post a crash landing in their Mi-220 Assault
Gunship "Fulcrum", attempting to survive the xenomorph outbreak. Whilst
successful in holding their ground despite their initial casualties, the
group are now running low on ammunition and their equipment is starting
to falter. Rescue is coming, but from whom? The Americans? Is the enemy
of my enemy really our friend?

---------------

Technically speaking the PR will see the deployment of the following
survivor roles
* 4 Riflemen
* 1 Non-Combat Synthetic
* 2 Sappers
* 2 Medics
* 1 Specialist
* 1 Squad Leader

![Screenshot 2023-07-10 01 54
11](https://github.com/cmss13-devs/cmss13/assets/6595389/07ef824b-ec62-499c-b7fd-80e935140b73)
Squad Patch

The UPP are in dire straits and this is represented in game by them
being spawned with a mish mash of armour and items. Only the Specialist
and SL are guaranteed to spawn with full armour, everyone else is likely
getting clothing or webbing. If a soldier spawns with webbing they will
be given a flashlight in their suit storage slot.

Weapon's wise it is a even 33% chance for each role to either get a UPP
Pulse Rifle, Pulse Carbine or a UPP Shotgun, with each only getting two
spare mags/3 shell handfuls. The Specialist is guaranteed to get the UPP
Pulse Rifle/Flamer. The mini-gun in the DS is a prop and non-functional.

The area is littered with mostly props or junk, the 'HEAP' ammo mags are
all empty. There is a single ammo box of soft-point ammo set to around
120 rounds.

Every UPP member speaks Russian by default, with Chinese and German as
their secondary languages. Only the SL and Synth can speak English.

The Synthetic has lesser equipment compared to its ERT counterpart and
uses survivor synth skills. It has similar combat restrictions to its
USCM counterpart.

The lore has been run past the lore channel and has been given a thumbs
up.

-----------
**Other Additions**

- External webbing can be clipped to UPP uniforms, which is used to
facilitate some of the cosmetic options for the survivors.
- New UPP logo has been added for documents and faxes. Official UPP
documents now can look more official (see images for example).
- UPP synthetics will now speak Russian as their first language instead
of English

-----------

The PR (as far as I can tell) works and runs properly. I mostly have a
few other tasks of moderate difficulty to finish.
To-Do
- [x] Add the UPP synth
- [x] Figure out how to get the UPP to spawn with random UPP names
- [x] Sprite the dropships side and rear wings
- [x] Expand the in-game lore and descriptions
- [x] Get spawn chances set to correct odds (20% ?)
- [ ] BALANCE

# Explain why it's good for the game

I believe that this would add further variety to the game in two
respects.

Firstly, the UPP are a somewhat under-used aspect that generally only
appear during events or as hostile ERT. Technically in-universe the USCM
and UPP are in a cold war and are technically not at war, I believe that
their introduction as a uncommon survivor spawn may provide some
interesting opportunities for roleplay to occur, either from actual
encounters between UPP and USCM personnel, or from the USCM finding the
last stand of the UPP.

Secondly I think more variety in survivor roles is a boon for the game,
and the opportunity to play as a UPP soldier in a non-ERT setting may
present more opportunities for UPP roleplay to occur.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![2023-07-19 15 51
28](https://github.com/cmss13-devs/cmss13/assets/6595389/ffbce5b8-e593-4f1e-b1c0-0035c56807fd)
In game map as it stands

![Screenshot 2023-07-09 16 10
32](https://github.com/cmss13-devs/cmss13/assets/6595389/76159e01-6d7f-44f7-b404-e30c54ecc9c8)
Overview of the UPP as they stand

![Screenshot 2023-07-10 01 54
11](https://github.com/cmss13-devs/cmss13/assets/6595389/07ef824b-ec62-499c-b7fd-80e935140b73)
Squad Patch

![Screenshot 2023-07-11 12 09
54](https://github.com/cmss13-devs/cmss13/assets/6595389/572ec5b1-55eb-4e4b-95d2-9957dd6ebb02)
Your Orders + UPP logo

</details>


# Changelog
:cl:
add: A team of UPP survivors may spawn on Trijent Dam. 
mapadd: A new nightmare insert for Trijent Dam has been added, featuring
a crashed UPP dropship.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>