fix: snowing inside walls on Sorokyne Strata map. (#4088)
# About the pull request
fixes: https://github.com/cmss13-devs/cmss13/issues/1496
change the area so it as a wall if this sort of wall is suppose to be
some sort of wall that snow is falling on sprite should reflect it..
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
it shouldn't snow inside walls? atleast i think it shouldn't...
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: changing an area so it stop snowing inside walls.
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>