Auxiliary Support Officer (#4081)
# About the pull request

This PR adds the Auxiliary Support Officer.

Read more here:
<https://docs.google.com/document/d/1aZLVaIwopVj_tant1Wcjuqo0_Ov70IL3U-UHfg7P7wo/edit?usp=sharing>

Main points:
Auxiliary Support Officer oversees the hangar, the intel team, the
engineering department, and the requisitions department.

Chief Engineer has been demoted to O1.

Requisitions Officer renamed to Quartermaster and is now a Gunnery
Sergeant.

For the duration of testing the following is true:

The Auxiliary Support Officer may not deploy. 

The Chief Engineer will be given explicit authority over construction
and control of the Forward Operating Base. The Chief Engineer, and
Maintenance Technicians under them, may deploy freely if there is an
Auxiliary Support Officer. If there is not an Auxiliary Support Officer
the Chief Engineer may deploy with permission from the Acting Commander
and Maintenance Technicians may deploy with the permission of the Chief
Engineer. The Chief Engineer is not required to assist with FOB
construction unless ordered to. One member of Engineering, or the
Auxiliary Support Officer, should be aboard the ship at all times.

The Quartermaster will have unchanged deployment standards as the
Requisitions Officer. Cargo Technicians may deploy if given permission
by the Quartermaster, Auxiliary Support Officer, or aCO. The Mess
Technician will be formally put under the Quartermaster in Requisitions.


# Explain why it's good for the game

Explained in above linked documentation.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow, Thwomper
add: Added Auxiliary Support Officer
add: Auxiliary Support Officer oversees the hangar, the intel team, the
engineering department, and the requisitions department
add: Chief Engineer has been demoted to O1.
add: Requisitions Officer is now "Quartermaster" and a Gunnery Sergeant.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>