Fix drop ships having inconsistent ceiling setting to prevent OBs (#4133)
# About the pull request

This PR is a followup to #3237 and in response to Round 18077 on New
Varadero which had a cluster OB wipe a large portion of the hive on the
dropship.

# Explain why it's good for the game

OB protections should not be inconsistent. If a dropship is supposed to
provide protection, then it should do so on all maps.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl: Drathek
fix: Fix new map dropship areas not getting changes to dropship ceilings
that prevents OBs
/:cl: