Fix remove iv needing power to operate. (#4070)
# About the pull request
fixes: https://github.com/cmss13-devs/cmss13/issues/3848
and add a skill check like for the blood pack.
also turned four variable with a single letter into better ones.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Add a skill check to operate IV.
fix: Fix remove iv needing power to operate.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>