SMG Gun+Mag Holster (#3825)
# About the pull request

Adds a new holster for the SMG. It can hold a M39 SMG plus two M39
magazines, it replaces the old SMG belt.

# Explain why it's good for the game

The M39 SMG is intended as an inferior firearm to the mainline M41A. The
perceived benefit of a M39 would be its lightness compared to the
bulkier M41A, however this is not reflected in-game outside of the M39
not confering a movement speed down when wielded, which is shared with
many other bulky guns like the M4RA.

The idea around this new holster is that it would give Marines a better
reason to consider taking the M39, as Marines in a more support oriented
role can access a unique holster that lets them carry additional
magazines in addition to the firearm itself.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2023-07-07 19 17
20](https://github.com/cmss13-devs/cmss13/assets/6595389/8d906b5b-cdd9-4c72-a5f2-9ada67264557)
Current iteration

![Screenshot 2023-07-07 01 58
02](https://github.com/cmss13-devs/cmss13/assets/6595389/8e5f4937-e780-427c-80ad-5b85d463b405)
![Screenshot 2023-07-07 02 01
46](https://github.com/cmss13-devs/cmss13/assets/6595389/64ad3a9e-f85c-448a-81b9-ca06d8fd8e3e)
Old iteration

![Screenshot 2023-07-06 17 17
56](https://github.com/cmss13-devs/cmss13/assets/6595389/15217e97-5a1a-4a8a-8bf4-1520d45355c8)
Old sprite and internal storage view

OBJ sprites and the inventory capacity as seen in game

</details>


# Changelog
:cl:
add: The SMG belt has been changed. It now holds a SMG plus two extra
magazines, plus it can fit in either the belt or armour slot.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>