updating MP and warden vendors. adding handgun case. (#4007)
# About the pull request
So the idea is to update the vendors for MP and warden because they are
outdated.
i replace the full belt by a handgun case choice as i found them more
pleasing to use.
i added a bunch of option for pouches and belt(including the previous
available handgun belt but empty) that seem fitting to me

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
It expand the option MP and Warden when they equip at start of the game.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Update the MP and Warden vendor to fallow more the standard of the
rifleman vendor.
add: change the selection of the handgun from a full belt to a handgun
case(3magazin,one railgunlight and the gun.).
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>