Fix runtime with hijack burrowed PR (#4144)
# About the pull request

This PR is a followup to #3592 fixing runtimes when hijack occurs
because it was assumed the mob's job was a datum/job when it actually is
a string.

# Explain why it's good for the game

Code shouldn't runtime.

Fixes: 

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/b9f47464-07d3-4972-af6d-21dd7dffd3e9)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fixed runtimes with hijack not calculating roles correctly for
hijack larva surge
/:cl: