Fix beams for IV drip machines when adding/removing a blood bag (#4145)
# About the pull request

This PR is a followup to #3645 fixing the beam getting into an incorrect
state when adding/removing a bag when the drip is already connected to
someone. Also fixes a lingering reference to the attached mob ever the
drip is destroyed.

# Explain why it's good for the game

Fixes this: 


https://github.com/cmss13-devs/cmss13/assets/76988376/77a69934-0eb9-4322-b0c5-74a53a34700f

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fix IV Drip machines not displaying the IV line correctly when
adding/removing a bag when already attached
code: Fixes a lingering reference to the mob when a IV drip machine is
destroyed.
/:cl: