autowiki port and gun autowiki (#4107)
MANY thanks to mothblocks of tg on
https://github.com/tgstation/tgstation/pull/64417 for this, it's very
cool!

this ports the autowiki (https://tgstation13.org/wiki/Guide_to_autowiki)
and can be used to generate wiki page templates instead of monotonously
updating damage values Forever!

this is a weird one to changelog so:

:cl: mothblocks, harry
add: added the backend functionality for autowiki, alongside automating
much of the work of maintaining guns on the wiki
/:cl: