Alters ASay chat color back to the old MSay color (#4151)
Changing Asay back to the Msay color that many were used to for a very
long time.

# Explain why it's good for the game

Because this color doesn't blend in with other major chat colors (such
as dsay, and others), and it reminds us of simpler times.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
admin: Adjustment of ASay color
/:cl:

---------

Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>
Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>
Co-authored-by: harryob <me@harryob.live>