Item Strip changes (#4153)
# About the pull request
Makes it so when stripping someone's items their skills are taken into
consideration. Each level of policing and CQC skills adds half a second
to the base delay. The skills are not taken into consideration if the
target is dead, unconscious or restrained.
Also adds an unstrippable trait, preventing items being forced onto or
removed from a mob. This also does not apply if dead, knocked out or
restrained.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
It makes a bit more sense for people trained in policing or CQC to be
better at resisting having their equipment stolen.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl:
add: Added a proc to calculate item strip delay, taking into account the
target's skills and adding 0.5s per level in Policing and CQC.
add: Added an unstrippable mob trait to prevent inventory manipulation.
Gave this trait to Working Joes.
/:cl: