Fix backpack sprayers (#4161)
# About the pull request
fixes: https://github.com/cmss13-devs/cmss13/issues/3369
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: spraying with the hose will no cost 5u of water while spaying 5u of
water.
fix: fixed the verb to take out the hose.
fix: remove the option to change the output as it was confusing for
people.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>