Foxtrot cryomarines spawn as freed mobs if not taken (#4132)
# About the pull request

Foxtrot cryomarines spawn as freed mobs if not taken.

# Explain why it's good for the game

It sucks when only 3 cryomarines spawn in because there is not enough
candidates. Now people can take cryomarines later.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
add: foxtrot cryomarines spawn as freed mobs if not taken.
/:cl:

---------

Co-authored-by: ihatethisengine <treml.treml@yandex.ru>
Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>