LV522 Maint Tunnel now useable by xenos as tunnels (#4186)
# About the pull request

This PR lets Xenos use the maintenance hatches scattered around on LV522
as tunnels and also removes the prop tunnel now that it has an actual
use so as to not confuse players if a prop one was added to a map

https://i.imgur.com/qwBobX1.png

# Explain why it's good for the game

In my mind, I'm attempting to draw from the movie moment where Hudson is
pulled down under the floor obviously that isn't exactly right but I
want to show that on maps where these new tunnel subtypes are placed,
Xenos have complete control of every aspect of the colony.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: SpartanBobby gdifirehawk
add: Subtype of tunnel added "Maintenance Tunnel" currently only on
LV522 these tunnels act the same as regular tunnels but look different,
keep an eye out marines. Description by GDIFIREHAWK
/:cl: