Fix solaris ridge map link and refactor current_map (#4192)
# About the pull request

This PR fixes the discrepancy between maps and a current_map counterpart
by removing the duplicate definitions. Additionally this will make
mapped in maps work for solaris ridge, soro has its name/desc set to the
latest edit, and the wiki example config is correct again.

# Explain why it's good for the game

Fixes #4191 and makes it harder for discrepancies like this to occur
again.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/e1651529-ca5d-4cb0-b198-be084ca70074)

</details>

# Changelog
:cl: Drathek
fix: Fixed mapped in maps for solaris ridge not opening correctly.
refactor: Refactored current_map to use a global list instead to remove
the need for duplicate definitions.
/:cl: